requirements.txt
python-dotenv==1.0.*
python-telegram-bot==13.5.*
urllib3==1.26.15

